refactor: :sparkles: Restructure CLI commands into separate modules and update entry point to 'envm'
